Update staging detection and README
Having a version in 'app.yaml' is no longer supported, instead
'gcloud preview app deploy' will generate a unique version id.
Changed the detection of "staging" to support that.

Removed most of the logging to stdout, to see if that has any
effect on the stability in production.

This is deployed as version 20150618t092953.

Updated README.md.

R=kustermann@google.com
BUG=

Review URL: https://codereview.chromium.org/1186993005.