Fix Analysis store: parent objects must exist.
Because both PackageAnalysis and PackageVersionAnalysis
stores the id of the Analysis, we either need to create
their instance first with null id (with all of the additional
error handling that comes with it), or generate the id
ourselves.

Using millis since epoch is low risk here, as Analysis id
is prefixed with both package and version.