Use global (reverse) package dependency graph to determine which packages need re-analysis (#388)
* Use global (reverse) package dependency graph to determine which packages need re-analysis

When startin up, the analysis service builds up a dependency graph
consisting of all packages in the database.  We track two different uses
of a package (as a direct or a dev dependency) via a special "dev/"
prefix.

The building of the initial dependency graph takes around 90 seconds
and consumes maybe around 20 MB.  Afterwards the dependency graph is
changed incrementally when new [PackageVersion]s are uploaded and is
used to answer the question of which packages need re-analysis.

Currently the implementation is a conservative estimate: We
say "A" needs re-analysis if a new [PackageVersion] of "B" is uploaded iff

  * any version of "A"'s pubspec.yaml file has "B" in the
    dependencies/dev_dependencies section

  * any version of "A"'s pubspec.yaml file has "X" in the dependencies
    section and "X" needs re-analysis if a new version of "B" is
    uploaded

This conservative estimate gives a superset of the actual set of
packages affected.  The current algorithm can be improved by taking
constraints into account.  Namely prevent the fixpoint algorithm to
including packages which have constraint violations.

It would also be possible to add the flutter sdk as another node in the
graph.  This would then also allow re-analysis of packages on new
flutter versions.  (Though currently we pinpoint the flutter version, so
it would make no difference at all).

Together with
[Create new Analysis only iff something changed](https://github.com/dart-lang/pub-dartlang-dart/issues/381)
it should allow us to get rid of re-analyzing all packages every X days
and the GC of analysis datastore entities (though we might still keep it
for other reasons).

* Move logic to frontend post-upload-hook

* Fix some mocks

* Remove batch support in notification

* run dartfmt