Increase pub_worker upload limits for blob.
Increase to 300MB, because we store dartdoc content twice:
 * As individually gzipped processed JSON files, and,
 * As a compressed tar.gz file.

For packages like `package:googleapis` the generated dartdoc is about
1.6 GB, this is a bit crazy, but it actually compresses down to 92 MB.
So it's not unreasonable to allow 300MB blobs from the pub_worker.

For the index, which pub.dev will load into memory when serving results,
we want to set a lower limit. We could tweak this in the future
for now we're limiting it to 64MB, just to have some sanity limit.